requirements.txt
black==24.3.0
flake8==7.0.0
python-dotenv==1.0.1 
